Move web to feature.web
