+++ app/src/main/java/com/olderwold/reddit/feature/web/ChromeTabWebPage.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.web
+package com.olderwold.reddit.feature.web
 
 import android.app.Activity
 import android.net.Uri
