+++ app/src/main/java/com/olderwold/reddit/feature/web/WebPageModule.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.web
+package com.olderwold.reddit.feature.web
 
 import android.app.Activity
 import dagger.Module
