+++ app/src/main/java/com/olderwold/reddit/feature/web/WebViewPage.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.web
+package com.olderwold.reddit.feature.web
 
 import android.app.Activity
 import android.content.ActivityNotFoundException
